Add generic throughput metric
This adds the ability for benchmarks to provide a getThroughputMetric
function, which returns a metric (e.g. GiB) that can be used to compute
throughput (e.g. GiB/s), as well as its unit string. The framework then
computes the maximum throughput using the minimum running time measured.

Additionally change the "comment" field of ResultConsumer::consumeResult
to be a "unit" instead, as that is the only thing we use it for.